+++ CHANGELOG.md
@@ -1,8 +1,8 @@
-#### 2026-2-26
+### 2026-2-26
 
 #### New:
 - 加入素数判断[`NumberTheory::IsPrime::miller_rabin`]
-
+- 加入矩阵快速幂[`Linear::qpow::matrixPos`]
 -----
 ### 2026-2-8